requirements.txt
gitpython
pysftp~=0.2.9
termcolor~=1.1.0
pytest
requests~=2.23.0